chore: config native_dialog_plus to support SPM
