+++ ios/.gitignore
@@ -36,5 +36,7 @@ Icon?
 /Flutter/Generated.xcconfig
 /Flutter/ephemeral/
 /Flutter/flutter_export_environment.sh
+/Flutter/ephmeral/
+.swiftpm/
 
 .build/
