--- ios/Classes/NativeDialogPlusPlugin.h
@@ -1,15 +1,0 @@
-#import <Flutter/Flutter.h>
-
-#import <Foundation/Foundation.h>
-
-NS_INLINE NSException * _Nullable tryBlock(void(^_Nonnull tryBlock)(void)) {
-@try {
-tryBlock();
-}
-@catch (NSException *exception) {
-return exception;
-}
-return nil;
-}
-@interface NativeDialogPlusPlugin : NSObject<FlutterPlugin>
-@end
