+++ ios/native_dialog_plus.podspec
@@ -4,7 +4,7 @@
 #
 Pod::Spec.new do |s|
   s.name             = 'native_dialog_plus'
-  s.version          = '0.0.1'
+  s.version          = '1.1.3'
   s.summary          = 'Minimalistic plugin to show native iOS Dialogs'
   s.description      = <<-DESC
 Minimalistic plugin to show native iOS Dialogs
@@ -13,9 +13,11 @@ Minimalistic plugin to show native iOS Dialogs
   s.license          = { :file => '../LICENSE' }
   s.author           = { 'Your Company' => 'email@example.com' }
   s.source           = { :path => '.' }
-  s.source_files = 'Classes/**/*'
+  s.source_files = 'native_dialog_plus/Sources/native_dialog_plus/**/*.swift'
+  s.exclude_files = 'native_dialog_plus/Sources/native_dialog_plus/NativeDialogPlusPlugin.{h,m}'
+  s.public_header_files = 'native_dialog_plus/Sources/native_dialog_plus/**/*.h'
   s.dependency 'Flutter'
-  s.platform = :ios, '9.0'
+  s.platform = :ios, '12.0'
 
   # Flutter.framework does not contain a i386 slice.
   s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES', 'EXCLUDED_ARCHS[sdk=iphonesimulator*]' => 'i386' }
