+++ ios/native_dialog_plus/Sources/native_dialog_plus/NativeDialogPlusPlugin.h
@@ -1,0 +1,6 @@
+#import <Flutter/Flutter.h>
+
+#import <Foundation/Foundation.h>
+
+@interface NativeDialogPlusPlugin : NSObject<FlutterPlugin>
+@end
